backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
channels==4.0.0
charset-normalizer==3.3.0
Django==4.2.5
django-nextjs==2.4.0
django-rest==0.8.7
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.1
django-environ==0.4.5
django-storages==1.11.1
django-cors-headers==3.7.0
django-braces==1.14.0
django-extensions==3.1.3
psycopg2-binary==2.8.6
arrow==1.1.0
gunicorn==20.1.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
websockets==11.0.3
yarl==1.9.2
